% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/little_and_su.R
\name{little_and_su}
\alias{little_and_su}
\title{Little and Su Imputation}
\usage{
little_and_su(
  df,
  window = NULL,
  id_col = NULL,
  imputation_classes_col = NULL,
  verbose = FALSE
)
}
\arguments{
\item{df}{data.frame, matrix, or tibble. Wide dataset with missing values. Each row should represent some unit and each column should represent a time period.
Besides this data an id column and an imputation classes column can be included. See id_col and imputation_classes_col for more.}

\item{window}{Odd integer. The window within which row effects are calculated. For panel data without many complete observations, consider
setting window to some odd integer. Defaults to NULL (which uses all time periods to calculate row effects).}

\item{id_col}{Integer or character. Identifies the column in the dataset with ID values for each observation. User may supply
either the number of the column or its name. Defaults to NULL.}

\item{imputation_classes_col}{Integer or character. Identifies the column in the dataset with imputation class values for each observation.
User may supply either the number of the column or its name. Defaults to NULL.}

\item{verbose}{Logical. If verbose is true, information about the imputation will be printed when little_and_su() is run.}
}
\description{
This function imputes missing values in a logitudinal dataset using the Little and Su algorithm (Little and Su 1989).
Each imputed missing value is based on some effect calculated by values from the same unit observed at a different time and
some effect calculated by other units' values in the missing value's time period.
}
